requirements.txt
git+https://github.com/Silverster98/pointops
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/pytorch3d.git@stable
wandb==0.14.0
tensorboard==2.12.1
loguru==0.6.0
hydra-core==1.3.2
matplotlib==3.7.1
tabulate==0.9.0
einops==0.6.1
natsort==8.3.1
pyquaternion==0.9.9
pandas==2.0.1
trimesh==3.21.7
pyrender==0.1.45
smplx==0.1.28
tqdm==4.65.0
smplkit==0.0.3
opencv-python==4.8.0.74
scikit-learn==1.2.2